pyproject.toml
[project]
name = "music-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.9.2",
    "django>=5.2.4",
    "django-debug-toolbar>=5.2.0",
    "django-stubs[compatible-mypy]>=5.2.2",
    "factory-boy>=3.3.3",
    "faker>=37.4.2",
    "mutagen>=1.47.0",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.1",
]


[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
cache-dir = "/tmp/ruff-cache/"
exclude = [
    "**migrations/",   
    "tests/",          
    "**/__pycache__/", 
    ".venv/",
    "**.mypy_cache",              

]
# Линтинг-настройки
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",      # Докстринги
    "D203",    # Форматирование докстрингов
    "D213",    # Многострочные докстринги
    "FA102",   # Запрещённые импорты
    "COM812",   # Конфликт с форматтером (добавлено в ignore)
    "RUF002",
    "T201",
    "S311",
    "ANN401",    # Использование Any
    "RUF012", # typing.ClassVar
   
]


[tool.ruff.lint.isort]
no-lines-before = [
    "standard-library",
    "local-folder"
]
known-third-party = []
known-local-folder = ["whole_app"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "ANN401",  # Динамические аннотации
    "S101",     # Использование assert
    "S311",     # Псевдослучайные значения
    
]


[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "music_stream.settings"
